remove proxy-deps image and fixup tests
