--- bin/docker-build-proxy-deps
@@ -1,21 +1,0 @@
-#!/bin/sh
-
-# Builds (or pulls) our proxy-deps docker image.
-
-set -eu
-
-if [ $# -ne 0 ]; then
-    echo "no arguments allowed for $(basename $0), given: $@" >&2
-    exit 64
-fi
-
-. bin/_docker.sh
-. bin/_tag.sh
-
-tag=$(proxy_deps_sha)
-
-if (docker_pull proxy-deps "${tag}"); then
-    echo "$(docker_repo proxy-deps):${tag}"
-else
-    docker_build proxy-deps "${tag}" proxy/Dockerfile-deps
-fi
