+++ bin/docker-images
@@ -22,4 +22,3 @@ docker_image cli          "${tag}"
 docker_image cli-bin      "${tag}"
 
 docker_image go-deps      "$(go_deps_sha)"
-docker_image proxy-deps   "$(proxy_deps_sha)"
