--- bin/update-proxy-deps-shas
@@ -1,11 +1,0 @@
-#!/bin/sh
-
-set -eu
-
-# Updates the tag for `runconduit/proxy-deps` across all Dockerfiles in this repository.
-
-sha=$(. bin/_tag.sh ; proxy_deps_sha)
-
-for f in $( grep -lR --include=Dockerfile\* proxy-deps: . ) ; do
-    sed -E -i'' -e "s|runconduit/proxy-deps:[^ ]+|runconduit/proxy-deps:${sha}|" "$f"
-done
