+++ proxy/Cargo.toml
@@ -50,3 +50,4 @@ libc = "0.2"
 
 [dev-dependencies]
 quickcheck = "0.6"
+conduit-grpc = { path = "./grpc", features = ["arbitrary"] }
