+++ proxy/grpc/Cargo.toml
@@ -4,6 +4,9 @@ version = "0.2.0"
 authors = ["Oliver Gould <ver@buoyant.io>"]
 publish = false
 
+[features]
+arbitrary = ["quickcheck"]
+
 [dependencies]
 convert = { path = "../convert" }
 
@@ -18,5 +21,9 @@ prost-types = "0.3.0"
 tower-grpc = { git = "https://github.com/tower-rs/tower-grpc" }
 tower-h2   = { git = "https://github.com/tower-rs/tower-h2" }
 
+quickcheck = { version = "0.6", optional = true }
+
+
 [build-dependencies]
 tower-grpc-build = { git = "https://github.com/tower-rs/tower-grpc" }
+
