+++ proxy/tests/support/mod.rs
@@ -1,7 +1,9 @@
 #![allow(unused)]
 
 extern crate bytes;
+pub extern crate conduit_grpc;
 extern crate conduit_proxy;
+pub extern crate convert;
 extern crate futures;
 extern crate h2;
 pub extern crate http;
